[refator][update] move single use components close to parent
