+++ src/components/index.jsx
@@ -2,11 +2,8 @@ export { Filters } from "./Filters";
 export { Footer } from "./Footer";
 export { Navbar } from "./Navbar";
 export { ScrollToTop } from "./ScrollToTop";
-export { CartCard } from "./CartCard";
-export { PriceSummary } from "./PriceSummary";
-export { ProductCard } from "./ProductCard";
+export { PriceSummary } from "../pages/cart/PriceSummary";
 export { EmptyPage } from "./EmptyPage";
 export { MockAPI } from "./MockAPI";
 export { Loader } from "./Loader";
-export { WishlistCard } from "./WishlistCard";
 export { InputError } from "./InputError";
