+++ src/pages/cart/Cart.jsx
@@ -1,9 +1,10 @@
-import React, { useEffect } from "react";
-import { Footer, Navbar, PriceSummary } from "../../components";
+import React from "react";
+import { Footer, Navbar } from "../../components";
 import { CartItems } from "./CartItems";
 import styles from "./Cart.module.css";
 import { useProduct } from "../../context/product-context";
 import { useDocumentTitle } from "../../utils";
+import { PriceSummary } from "./PriceSummary";
 
 export function Cart() {
   const {
